Combine io and io_dng_sdk
